updates blog filter backend and separate project table for maintainability
