+++ frontend/src/components/blog/blog-carousel.tsx
@@ -1,4 +1,3 @@
-import { FormButton } from "../components/form/button";
 export function BlogCarousel() {
   return (
     <>
